05_Our_First_Agent_with_LangGraph/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
langsmith==0.0.69
tavily-python==0.3.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
arxiv==2.1.0
requests==2.31.0
